Feature|UIUX: Created Basic Add Panel
Created a basic add panel that opens on button click. This panel has an add button that when clicked will add another circle to the wrap panel. I have included a scroll bar that appears when too many objects have been added. I will need to expand this in the next commit, I will be making a new branching to try add images.